requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.9.2
numpy>=1.23.0
scipy>=1.9.0
pandas>=1.5.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.12.0
soundfile>=0.12.1
noisereduce>=2.0.0
tqdm>=4.64.0
pyyaml>=6.0
streamlit>=1.18.0
python-multipart>=0.0.5
typing-extensions>=4.4.0
requests>=2.28.0
